// src/clients/models/person.ts
/* eslint-disable @typescript-eslint/naming-convention */
export type Person = {
	name: string;
	height: string;
	mass: string;
	hair_color: string;
	skin_color: string;
	eye_color: string;
	birth_year: string;
	gender: string;
	homeworld: string;
	films: string[];
	species: string[];
	vehicles: string[];
	starships: string[];
	url: string;
	created: string;
	edited: string;
};